refactor(utils): move stylize.py to widgets/
stylize.py provides wx widget styling helpers (stylize_button, stylize_choice,
stylize_label, stylize_textctrl) and is imported only from 10 sites under
widgets/ (buttons, panels). It belongs in the widget layer, not utils.

Refs #441
