+++ widgets/panels/sideboard_guide_panel/frame.py
@@ -19,7 +19,7 @@ from utils.constants.ui_layout import (
     GUIDE_COL_NOTES_WIDTH,
     PADDING_MD,
 )
-from utils.stylize import stylize_button
+from widgets.stylize import stylize_button
 from widgets.panels.sideboard_guide_panel.handlers import SideboardGuidePanelHandlersMixin
 from widgets.panels.sideboard_guide_panel.properties import SideboardGuidePanelPropertiesMixin
